backport to 1.21.5

+++ common/src/main/java/org/vivecraft/client/extensions/BufferBuilderExtension.java
@@ -13,5 +13,5 @@ public interface BufferBuilderExtension {
     /**
      * @return size of the underlying buffer
      */
-    long vivecraft$getBufferSize();
+    int vivecraft$getBufferSize();
 }
